refactor(migration): move SpdxSbom Document impl to central location
Move the SpdxSbom type and its Document impl from the migration file
to migration/src/data/document/sbom.rs so future data migrations can
reuse the SPDX-filtered query without duplicating the type.

Implements TC-5397
