+++ migration/src/data/document/sbom.rs
@@ -1,7 +1,8 @@
 use super::Document;
 use bytes::Bytes;
-use sea_orm::{FromQueryResult, QuerySelect, prelude::*};
-use trustify_entity::sbom;
+use sea_orm::sea_query::{Expr, extension::postgres::PgExpr};
+use sea_orm::{FromQueryResult, QueryFilter, QuerySelect, prelude::*};
+use trustify_entity::{labels::Labels, sbom};
 use trustify_module_storage::service::StorageBackend;
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq, Hash, FromQueryResult)]
@@ -31,6 +32,38 @@ impl Document for Sbom {
 
     async fn all<C: ConnectionTrait>(tx: &C) -> Result<Vec<Self::Id>, DbErr> {
         sbom::Entity::find()
+            .select_only()
+            .column_as(sbom::Column::SourceDocumentId, "source")
+            .column_as(sbom::Column::SbomId, "sbom")
+            .into_model()
+            .all(tx)
+            .await
+    }
+
+    async fn source<S, C>(id: &Self::Id, storage: &S, tx: &C) -> Result<Self, anyhow::Error>
+    where
+        S: StorageBackend + Send + Sync,
+        C: ConnectionTrait,
+    {
+        super::load(id.source, storage, tx).await
+    }
+}
+
+/// SPDX-only Document variant that filters `all()` to SBOMs labeled `type=spdx`.
+pub struct SpdxSbom(pub Sbom);
+
+impl From<Bytes> for SpdxSbom {
+    fn from(value: Bytes) -> Self {
+        SpdxSbom(Sbom::from(value))
+    }
+}
+
+impl Document for SpdxSbom {
+    type Id = Id;
+
+    async fn all<C: ConnectionTrait>(tx: &C) -> Result<Vec<Self::Id>, DbErr> {
+        sbom::Entity::find()
+            .filter(Expr::col(sbom::Column::Labels).contains(Labels::new().add("type", "spdx")))
             .select_only()
             .column_as(sbom::Column::SourceDocumentId, "source")
             .column_as(sbom::Column::SbomId, "sbom")
